% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{arg_character}
\alias{arg_character}
\title{checks if correct datatype is provided to function (character check)}
\usage{
arg_character(name)
}
\arguments{
\item{name}{argument from function.}
}
\description{
checks if correct datatype is provided to function (character check)
}
\author{
Sergej Ruff
}
\keyword{internal}
