% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{arg_logical}
\alias{arg_logical}
\title{checks if correct datatype is provided to function (logical check)}
\usage{
arg_logical(name)
}
\arguments{
\item{name}{argument from function.}
}
\description{
checks if correct datatype is provided to function (logical check)
}
\keyword{internal}
