% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{arg_numeric_or_integer}
\alias{arg_numeric_or_integer}
\title{checks if correct datatype is provided to function (numeric check)}
\usage{
arg_numeric_or_integer(name)
}
\arguments{
\item{name}{argument from function.}
}
\description{
checks if correct datatype is provided to function (numeric check)
}
\details{
currently not in use. Might be in a later version.
}
\keyword{internal}
