% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxp_summary_stats.R
\name{boxp_summary_stats}
\alias{boxp_summary_stats}
\title{boxp_summary_stats:  Calculate summary statistics for boxplot}
\usage{
boxp_summary_stats(vh_file, group = "best_query", ycol = "ViralRefSeq_ident")
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter hittables.}

\item{group}{column to group stats by}

\item{ycol}{column, which specifies identity or evalue.}
}
\value{
A data frame containing summary statistics including median,
Q1 (1st quartile), Q3 (3rd quartile), minimum, and maximum values for each virus group
}
\description{
This function calculates summary statistics for boxplots created by vhgBoxplot function.
}
\details{
This function is an internal utility function used within the package.
It calculates summary statistics such as median, quartiles, minimum,
and maximum identity values for each virus group from the input dataset.
}
\author{
Sergej Ruff
}
\keyword{internal}
