% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{check_columns}
\alias{check_columns}
\title{internal: check if all columns exist}
\usage{
check_columns(file, columns)
}
\arguments{
\item{file}{file}

\item{columns}{single column or list of columns}
}
\description{
internal: check if all columns exist
}
\keyword{internal}
