% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{check_input_type}
\alias{check_input_type}
\title{internal: check if datatype is correct}
\usage{
check_input_type(vh_file, columns, option)
}
\arguments{
\item{vh_file}{file}

\item{columns}{object single or list specifying column}

\item{option}{1 for chacter,2 for numeric}
}
\description{
internal: check if datatype is correct
}
\keyword{internal}
