% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{check_is_dataframe}
\alias{check_is_dataframe}
\title{check if obj is a dataframe}
\usage{
check_is_dataframe(obj)
}
\arguments{
\item{obj}{obj}
}
\value{
An error if obj is not a dataframe
}
\description{
check if obj is a dataframe
}
\author{
Sergej Ruff
}
\keyword{internal}
