% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{colorbildsupport_}
\alias{colorbildsupport_}
\title{internal: add colorblind support by using viridis color palettes for identity scatterplot}
\usage{
colorbildsupport_(plot, option)
}
\arguments{
\item{plot}{plot obj.}

\item{option}{A character string indicating the color map option to use. Eight options are available:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E")
\item "rocket" (or "F")
\item "mako" (or "G")
\item "turbo" (or "H")
}}
}
\value{
plot obj
}
\description{
internal: add colorblind support by using viridis color palettes for identity scatterplot
}
\keyword{internal}
