% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{colorblind_support}
\alias{colorblind_support}
\title{Internal function: adds colour blind support}
\usage{
colorblind_support(plot, color)
}
\arguments{
\item{plot}{plot obj}

\item{color}{color}
}
\value{
plot
}
\description{
Internal function: adds colour blind support
}
\keyword{internal}
