% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineHittables.R
\name{CombineHittables}
\alias{CombineHittables}
\title{CombineHittables: Combine hittables.}
\usage{
CombineHittables(...)
}
\arguments{
\item{...}{Hittables to be combined.}
}
\value{
A single hittable resulting from the row-binding of all input Hittables.
}
\description{
CombineHittables combines multiple hittables by row-binding them, provided they have the same column names.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
file <- ImportVirusTable(path)
file2 <- ImportVirusTable(path)  # both files have 180 observations

combined_file <- CombineHittables(file,file2)

print(nrow(combined_file))
}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
