% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_table_eval_box.R
\name{creat_table_eval_box}
\alias{creat_table_eval_box}
\title{vh_refeval_iden_boxplot: Create a formatted table using the gt package}
\usage{
creat_table_eval_box(summary_stats)
}
\arguments{
\item{summary_stats}{A data frame containing summary statistics}
}
\value{
A formatted gt table
}
\description{
This function takes summary statistics and creates a formatted table using
the gt package.
}
\details{
This function is for internal use within the package.
It formats summary statistics into a table
using the gt package for further evaluation.
}
\seealso{
\code{\link{vh_refeval_iden_boxplot}}
}
\keyword{internal}
