% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatTableIdentityBox.R
\name{creatTableIdentityBox}
\alias{creatTableIdentityBox}
\title{creatTableIdentityBox: Create a formatted table using the gt package for vhEvalIdenBoxplot
with eval_vs_iden="identity"}
\usage{
creatTableIdentityBox(summary_stats_identity)
}
\arguments{
\item{summary_stats_identity}{A data frame containing summary statistics}
}
\value{
A gt table representing the summary statistics
}
\description{
This function takes the summary statistics output from vhEvalIdenBoxplot
and creates a formatted table using the gt package.Should only be used if the output stats were
generated with eval_vs_iden="identity".
}
\details{
This function generates a table for the summary statistics output from the vhEvalIdenBoxplot
function.It formats summary statistics into a table using the gt package.
}
\seealso{
\code{\link{vhEvalIdenBoxplot}}
}
