% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{Current_ICTV}
\alias{Current_ICTV}
\title{Current_ICTV: Verify the version of ICTV data currently in use.}
\usage{
Current_ICTV()
}
\value{
data frame with the currently used ICTV data
}
\description{
Displays the entire ICTV data object currently used by Virusparies.
}
\details{
Virusparies utilizes the ICTV taxonomy data set for two main purposes:
\enumerate{
\item Assigning the appropriate taxonomy rank via VhgPreprocessTaxa and related functions.
\item In plotting functions, the ICTV data set is used to assign colors based on phylum.
}

Current_ICTV allows users to inspect the current ICTV data object used by Virusparies.
It also enables users to check if their custom version of the ICTV data,
provided by the \code{\link{New_ICTV}}, is loaded.
}
\examples{

internal_ICTV <- Current_ICTV()

str(internal_ICTV)


}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
