% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVirusGt.R
\name{exportVirusGt}
\alias{exportVirusGt}
\title{Export Graphical Tables}
\usage{
exportVirusGt(
  gtable,
  filename = "table.png",
  path = NULL,
  create.dir = FALSE,
  ...
)
}
\arguments{
\item{gtable}{A graphical table object.}

\item{filename}{Name of the output file. Default is "table.png".}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the
fully qualified file name. Defaults to the working directory.}

\item{create.dir}{Whether to create new directories if a non-existing directory is specified in
the filename or path (TRUE) or return an error (FALSE, default). If FALSE and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}

\item{...}{All other options passed to the appropriate internal saving function.}
}
\description{
This function allows the user to export graphical tables in different formats.
}
\details{
export graphical tables (gt) generated by functions within the Virusparies package.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)



}
\author{
Sergej Ruff
}
