% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVirusPlot.R
\name{exportVirusPlot}
\alias{exportVirusPlot}
\title{Export Virusparies Plots}
\usage{
exportVirusPlot(
  plot_object,
  file_name = "plot.png",
  width = 800,
  height = 600,
  units = "px",
  pointsize = 12
)
}
\arguments{
\item{plot_object}{A plot object to be exported.}

\item{file_name}{Name of the output file. Default is "plot.png".}

\item{width}{The width of the output file. Default is 800 pixels.}

\item{height}{The height of the output file. Default is 600 pixels.}

\item{units}{The units of the width and height parameters. Can be one of "in", "cm", "mm", or "px". Default is "px".}

\item{pointsize}{The point size of the font. Default is 12.}
}
\description{
This function allows the user to export plots in different formats.
}
\details{
export plots generated by functions within the Virusparies package.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

}
