% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{extract_taxa}
\alias{extract_taxa}
\title{internal: extract viridae element}
\usage{
extract_taxa(sublist, taxa_rank)
}
\arguments{
\item{sublist}{taxonomy column}
}
\description{
internal: extract viridae element
}
\keyword{internal}
