% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{extract_viridae}
\alias{extract_viridae}
\title{internal: extract viridae element}
\usage{
extract_viridae(sublist)
}
\arguments{
\item{sublist}{taxonomy column}
}
\description{
internal: extract viridae element
}
\keyword{internal}
