% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{facet_plot}
\alias{facet_plot}
\title{facet_plot}
\usage{
facet_plot(plot, facet_ncol = FALSE, flip_coords = TRUE)
}
\arguments{
\item{plot}{plot object}

\item{facet_ncol}{number of columns to facet by. Default NULL}

\item{flip_coords}{flip_coords}
}
\value{
plot
}
\description{
facet_plot
}
\keyword{internal}
