% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{filter_specific_group}
\alias{filter_specific_group}
\title{internal function filter_specific_group}
\usage{
filter_specific_group(file, groupby, filter_group_criteria)
}
\arguments{
\item{file}{a hittable file}

\item{groupby}{groupby column}

\item{filter_group_criteria}{string,number or vector of character or number to filte by.}
}
\value{
preprocessed hittable
}
\description{
internal function filter_specific_group
}
\keyword{internal}
