% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outlier_eval_box.R
\name{find_outlier_eval_box}
\alias{find_outlier_eval_box}
\title{find_outlier_eval_box: Detect outliers in a dataset using the interquartile range (IQR) method}
\usage{
find_outlier_eval_box(vh_file, group = "best_query")
}
\arguments{
\item{vh_file}{A numeric vector or column of a data frame}

\item{group}{column to group stats by}
}
\value{
A dataframe indicating outliers in the input data
}
\description{
This function detects outliers in a dataset using the interquartile range (IQR) method.
}
\details{
This function is an internal utility function used within the package.
It identifies outliers in a numeric vector or column of a data frame by comparing the values
to the lower and upper bounds determined by the interquartile range (IQR) method.
}
\keyword{internal}
