% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_run_bar_table.R
\name{generate_run_bar_table}
\alias{generate_run_bar_table}
\title{Generate a summary table of unique SRA runs for each virus group}
\usage{
generate_run_bar_table(vh_file)
}
\arguments{
\item{vh_file}{A data frame containing the Virushunter Hittables Results}
}
\value{
A formatted gt table summarizing unique runs for each virus group
}
\description{
This function generates a summary table of unique SRA runs for each virus group
based on the input dataset.
}
\details{
This function is an internal utility function used within the package.
It calculates the number of unique runs for each virus group from the input dataset.
}
\seealso{
\code{\link{vh_runs_bar}}
}
\keyword{internal}
