% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{get_plot_parameters}
\alias{get_plot_parameters}
\title{Internal function to set default title and cutoff in boxplot function}
\usage{
get_plot_parameters(y_column, cut)
}
\arguments{
\item{y_column}{y-column}

\item{cut}{cutoff value}
}
\value{
list containing cutoff and title
}
\description{
Internal function to set default title and cutoff in boxplot function
}
\keyword{interlal}
