% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{has_columnnames}
\alias{has_columnnames}
\title{check if dataframe has column names.}
\usage{
has_columnnames(df)
}
\arguments{
\item{df}{a dataframe obj}
}
\value{
error message, if df has no column names
}
\description{
check if dataframe has column names.
}
\author{
Sergej Ruff
}
\keyword{internal}
