% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVirusTable.R
\name{importVirusTable}
\alias{importVirusTable}
\title{importVirusTable: Import VirusHunterGatherer hittables into R}
\usage{
importVirusTable(path)
}
\arguments{
\item{path}{A character string specifying the path to the VirusHunter or VirusGatherer hittable file.}
}
\value{
A dataframe containing the data from the VirusHunter or VirusGatherer hittable file.
}
\description{
This function imports VirusHunterGatherer hittables into R.
}
