% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{is_vh_file_empty}
\alias{is_vh_file_empty}
\title{Check if hittable is empty}
\usage{
is_vh_file_empty(vh_file)
}
\arguments{
\item{vh_file}{a Virushuntegatherer hittable}
}
\description{
Check if hittable is empty
}
\details{
Internal helper function which checks if hittables are empty. Returns an error message if they are.
}
\keyword{internal}
