% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{New_ICTV}
\alias{New_ICTV}
\title{New_ICTV: Assign Custom ICTV Data for Use in Virusparies}
\usage{
New_ICTV(new_ICTV_data = NULL)
}
\arguments{
\item{new_ICTV_data}{data frame containing new ICTV data.}
}
\value{
An \code{ICTV_env} object containing the user-defined ICTV data
}
\description{
New_ICTV enables users to define their own ICTV data for use in Virusparies.
}
\details{
Virusparies utilizes the ICTV taxonomy data set for two main purposes:
\enumerate{
\item Assigning the appropriate taxonomy rank via VhgPreprocessTaxa and related functions.
\item In plotting functions, the ICTV data set is used to assign colors based on phylum.
}

\code{New_ICTV} enables users to define and load their own ICTV data for use in Virusparies.
This feature is useful for those who wish to use a subset of ICTV data to speed up functions
such as \code{\link{VhgPreprocessTaxa}}, or for users who want to incorporate a newer or older version of
ICTV data within Virusparies.
The custom ICTV data is stored in a new environment, which means it does not overwrite the
internal data. Instead, users need to apply this function each time they start a new R session
to load their custom dataset.
}
\examples{
# Define example data

# Sample data
Example_ICTV <- data.frame(
Phylum = c("Taleaviricota", "Taleaviricota", "Taleaviricota", "Taleaviricota",
"Taleaviricota"),Subphylum = c(NA, NA, NA, NA, NA),
Class = c("Tokiviricetes", "Tokiviricetes", "Tokiviricetes", "Tokiviricetes",
"Tokiviricetes"),
Subclass = c(NA, NA, NA, NA, NA),
Order = c("Ligamenvirales", "Ligamenvirales", "Ligamenvirales", "Ligamenvirales",
"Ligamenvirales"),
Suborder = c(NA, NA, NA, NA, NA),
Family = c("Lipothrixviridae", "Lipothrixviridae", "Lipothrixviridae",
"Lipothrixviridae", "Lipothrixviridae"),
Subfamily = c(NA, NA, NA, NA, NA),
Genus = c("Alphalipothrixvirus", "Alphalipothrixvirus",
"Betalipothrixvirus", "Betalipothrixvirus", "Betalipothrixvirus"),
Subgenus = c(NA, NA, NA, NA, NA),
stringsAsFactors = FALSE)

\donttest{
# Assign new example ICTV data for use in Virusparies
New_ICTV(Example_ICTV)
# check currently used ICTV data in Virusparies
Current_ICTV()

}




}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
