% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{plot_boxplot_message}
\alias{plot_boxplot_message}
\title{internal function: message to be printed in boxplot depending on chosen y-column}
\usage{
plot_boxplot_message(y_column, x_column, cutoff)
}
\arguments{
\item{y_column}{y_column}

\item{x_column}{x_column}

\item{cutoff}{cutoff}
}
\description{
internal function: message to be printed in boxplot depending on chosen y-column
}
\keyword{internal}
