% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_runs_bar.R
\name{preprocess_runs_bar}
\alias{preprocess_runs_bar}
\title{Preprocess virus data for generating a bar plot}
\usage{
preprocess_runs_bar(vh_file)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter Hittables results.}
}
\value{
A processed data frame for run_bar plot.
}
\description{
This function preprocesses virus data for generating a bar plot
by calculating the number of unique SRA runs for each virus group and
creating additional variables for visualization.
}
\details{
This function is an internal utility function used within the package.
It calculates the number of unique runs for each virus group from the VirusHunters Hittables.
}
\author{
Sergej Ruff
}
\keyword{internal}
