% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{remove_group_text}
\alias{remove_group_text}
\title{remove group text}
\usage{
remove_group_text(plot, remove_x_axis_labels, flip_coords)
}
\arguments{
\item{plot}{plot}

\item{remove_x_axis_labels}{remove_x_axis_labels}

\item{flip_coords}{flip_coords}
}
\description{
remove group text
}
\keyword{internal}
