% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{remove_samples}
\alias{remove_samples}
\title{remove sample(s) elements in taxonomy columns}
\usage{
remove_samples(sublist)
}
\arguments{
\item{sublist}{ViralRefSeq_taxonomy column from Hittable file}
}
\value{
list with preprocessed ViralRefSeq_taxonomy elements
}
\description{
remove sample(s) elements in taxonomy columns
}
\details{
Internal hellper function used in 'taxonomy_group_preprocess' function.
The ViralRefSeq_taxonomy column of Hittables sometimes contains entries such as "environmental sample"
as the second element. We want to remove those "sample" elements since we want to use the virus name in the second
element as the x_column/groupby instead of best_query. That way the third element becomes the second one.
}
\keyword{internal}
