% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_theme.R
\name{select_theme}
\alias{select_theme}
\title{select_theme}
\usage{
select_theme(theme_choice)
}
\arguments{
\item{theme_choice}{A character indicating which theme to use. Options include "minimal", "classic", "light", "dark", and "void".}
}
\value{
A ggplot2 theme object to be used in plots.
}
\description{
This function selects different themes for plots.
}
\details{
This function is an internal utility function used within the package.
It selects the theme to use for the different plots in this package.
}
\author{
Sergej Ruff
}
\keyword{internal}
