% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeViralStats.R
\name{SummarizeViralStats}
\alias{SummarizeViralStats}
\title{SummarizeViralStats: Generate summary stats outside of plot functions}
\usage{
SummarizeViralStats(
  file,
  groupby = "best_query",
  metric,
  metric_cutoff,
  filter_cutoff = NULL,
  show_total = FALSE,
  extra_stats = NULL
)
}
\arguments{
\item{file}{VirusHunterGatherer hittable.}

\item{groupby}{(optional): A character specifying the column containing the groups (default: "best_query").
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{metric}{A character string specifying the name of the metric column to be used for calculations.
This column must be present in \code{file}. Supported metric columns include:
\itemize{
\item "contig_len"
\item "ViralRefSeq_E"
\item "ViralRefSeq_ident"
}}

\item{metric_cutoff}{A numeric value used to classify the metric into two categories: below cutoff and above or equal to cutoff.}

\item{filter_cutoff}{A numeric value for optional filtering of the data based on E-value. Rows where the specified filtering column has a value
less than this cutoff are retained. If \code{NULL}, no filtering is applied. Default is \code{NULL}.}

\item{show_total}{A logical value indicating whether to include a row with the total sums for each column in the summary table.
Default is \code{FALSE}.}

\item{extra_stats}{A character vector specifying additional summary statistics to include in the output. Options include:
\itemize{
\item \code{"mean"}
\item \code{"median"}
\item \code{"Q1"}
\item \code{"Q3"}
\item \code{"sd"}
\item \code{"min"}
\item \code{"max"}
}
If \code{NULL} (the default), only the basic counts are included.}
}
\value{
A data frame summarizing the viral stats. The output includes:
\itemize{
\item The count of observations below and above or equal to the \code{metric_cutoff}.
\item Optional additional summary statistics as specified by \code{extra_stats}.
\item An optional total row if \code{show_total} is \code{TRUE}.
}
}
\description{
Summarizes data by grouping it according to a specified metric (contig length, E-value or Identity).
SummarizeViralStats generates a summary table that includes counts of observations based on a specified metric cutoff.
It computes relevant summary statistics depending on the selected metric, with options to filter rows
based on a cutoff value.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
file <- ImportVirusTable(path)

stats <- SummarizeViralStats(file=file,
groupby = "best_query",
metric = "ViralRefSeq_ident",
metric_cutoff = 90,
show_total = TRUE,
filter_cutoff = 1e-5,
extra_stats = c("median","Q1","Q3"))

print(stats)



}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
