% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{taxonomy_group_preprocess}
\alias{taxonomy_group_preprocess}
\title{preprocess ViralRefSeq_taxonomy elements}
\usage{
taxonomy_group_preprocess(vh_file)
}
\arguments{
\item{vh_file}{hittables file}
}
\value{
vh_file with preprocessed ViralRefSeq_taxonomy elements
}
\description{
preprocess ViralRefSeq_taxonomy elements
}
\details{
Besides best_query the user can utilize the ViralRefSeq_taxonomy column as x_column or groupby
in plots. That columns needs to be preprocessed as it it too long and has too many unique elements
to be used for grouping. The element containing "viridae" is used because the first one is the tax ID.
NA are replaced by "unclassified" and existing "unclassified" are removed.
}
\keyword{internal}
