% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{taxonomy_rank_hierarchy}
\alias{taxonomy_rank_hierarchy}
\title{define taxa string}
\usage{
taxonomy_rank_hierarchy(taxa_rank)
}
\arguments{
\item{taxa_rank}{a character indicating the taxa rank}
}
\value{
a character with the taxa suffix
}
\description{
define taxa string
}
\keyword{internal}
