% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VgConLenViolin.R
\name{VgConLenViolin}
\alias{VgConLenViolin}
\title{VgConLenViolin: Generate a Violinplot of contig length for each group (Gatherer only)}
\usage{
VgConLenViolin(
  vg_file = vg_file,
  taxa_rank = "Family",
  cut = 1e-05,
  log10_scale = TRUE,
  reorder_criteria = "median",
  adjust_bw = 1,
  jitter_point = FALSE,
  theme_choice = "linedraw",
  flip_coords = TRUE,
  title = "Violinplot of contig length for each group",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = NULL,
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Viral group",
  ylabel = "Contig length (nt)",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  remove_group_labels = FALSE,
  legend_title = "Phylum",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  min_observations = 1,
  facet_ncol = NULL,
  add_boxplot = FALSE
)
}
\arguments{
\item{vg_file}{A data frame containing VirusGatherer Hittable results.}

\item{taxa_rank}{(optional): Specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{cut}{(optional): The significance cutoff value for E-values (default: 1e-5).}

\item{log10_scale}{(optinal) transform y-axis to log10 scale (default: TRUE)}

\item{reorder_criteria}{Character string specifying the criteria for reordering the x-axis ('max', 'min', 'median'(Default),'mean').
NULL sorts alphabetically.}

\item{adjust_bw}{(optional): control the bandwidth of the kernel density estimator used to create the violin plot.
A higher value results in a smoother plot by increasing the bandwidth, while a lower value can make the plot more detailed but potentially noisier.
Default is 1.}

\item{jitter_point}{(optional): logical: TRUE to show all observations, FALSE to show only groups with less than 2 observations.}

\item{theme_choice}{(optional): A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "linedraw".}

\item{flip_coords}{(optional): Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional): A character specifying the title of the plot. Default title is set based on y_column.}

\item{title_size}{(optional): Numeric specifying the size of the title text. Default is 16.}

\item{title_face}{(optional): A character specifying the font face for the title text. Default is "bold".}

\item{title_colour}{(optional): A character specifying the color for the title text. Default is "#2a475e".}

\item{subtitle}{(optional): A character specifying the subtitle of the plot. Default subtitle is set based on y_column.}

\item{subtitle_size}{(optional): Numeric specifying the size of the subtitle text. Default is 12.}

\item{subtitle_face}{(optional): A character specifying the font face for the subtitle text. Default is "bold".}

\item{subtitle_colour}{(optional): A character specifying the color for the subtitle text. Default is "#1b2838".}

\item{xlabel}{(optional): A character specifying the label for the x-axis. Default is "Virus found in query".}

\item{ylabel}{(optional): A character specifying the label for the y-axis. Default is set based on y_column.}

\item{axis_title_size}{(optional): Numeric specifying the size of the axis title text. Default is 12.}

\item{xtext_size}{(optional): Numeric specifying the size of the x-axis tick labels. Default is 10.}

\item{ytext_size}{(optional): Numeric specifying the size of the y-axis tick labels. Default is 10.}

\item{remove_group_labels}{(optional): If \code{TRUE}, the group labels will be removed; if \code{FALSE} or omitted, the labels will be displayed.}

\item{legend_title}{(optional): A character specifying the title for the legend. Default is "Phylum".}

\item{legend_position}{(optional): A character specifying the position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional): Numeric specifying the size of the legend title text. Default is 12.}

\item{legend_title_face}{(optional): A character specifying the font face for the legend title text. Default is "bold".}

\item{legend_text_size}{(optional): Numeric specifying the size of the legend text. Default is 10.}

\item{min_observations}{(optional): Minimum number of observations required per group to be included in the plot (default: 1).}

\item{facet_ncol}{(optional):  The number of columns for faceting. Default is NULL.
It is recommended to specify this when the number of viral groups is high, to ensure they fit well in one plot.}

\item{add_boxplot}{add a boxplot to the violin plot (default: FALSE)}
}
\value{
A list containing the following components:
\itemize{
\item plot: A plot object representing the violin plot.
\item contiglen_stats: A tibble data frame with summary statistics for "contig_len" values.
}
}
\description{
This function creates a violin plot to visualize the distribution of contig lengths
for each group in VirusGatherer hittables.
}
\details{
This function creates a violin plot to visualize the distribution of contig lengths
for each group in the "ViralRefSeq_taxonomy" column of the VirusGatherer hittable.
The x-axis represents the groups as defined by the "ViralRefSeq_taxonomy" column,
and the y-axis represents the contig lengths.

By default, the y-axis is transformed to a log10 scale to better visualize differences
in contig lengths across groups. This transformation can be disabled by setting the
\code{log10_scale} argument to FALSE.

\code{min_observations} filters the datasets to include only groups with at least the specified number of observations
before plotting them. This feature allows users to exclude groups with insufficient data.
By default, every group is plotted, as the minimum requirement is set to at least one observation per group.
}
\examples{

# import gatherer files
path2 <- system.file("extdata", "virusgatherer.tsv", package = "Virusparies")
vg_file <- ImportVirusTable(path2)

# create a violinplot.
violinplot <- VgConLenViolin(vg_file=vg_file,cut = 1e-5,log10_scale = TRUE)

violinplot


}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
