% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_refeval_iden_boxplot.R
\name{vh_refeval_iden_boxplot}
\alias{vh_refeval_iden_boxplot}
\title{Generate boxplots comparing E-values or identity for each virus group}
\usage{
vh_refeval_iden_boxplot(vh_file, cut = 1e-05, eval_vs_iden = "evalue")
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter Hittable results.}

\item{cut}{The significance cutoff value for E-values (default: 1e-5)}

\item{eval_vs_iden}{Specifies whether to generate boxplots for E-values ("evalue")
or identity ("identity")}
}
\value{
A list containing the generated boxplot, summary statistics, and optionally outliers and summary tables
}
\description{
This function generates boxplots comparing either E-values or identity
for each virus group from VirusHunter Hittables results.
}
\details{
This function generates boxplots comparing either E-values or identity for each virus
group from the VirusHunters Hittable.
It also calculates summary statistics and optionally identifies outliers for further analysis.
The user can specify whether to generate boxplots for E-values or identity by setting
the parameter \code{eval_vs_iden}.
}
