% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_runs_bar.R
\name{vh_runs_bar}
\alias{vh_runs_bar}
\title{Generate a bar plot showing the number of datasets with unique runs found for each Virus group.}
\usage{
vh_runs_bar(vh_file, generate_table = FALSE)
}
\arguments{
\item{vh_file}{Dataframe containing Virushunter hittables results}

\item{generate_table}{Logical indicating whether to generate a table along with the plot
(default: FALSE)}
}
\value{
A list containing the bar plot and optionally the generated table and processed data
}
\description{
Generate a bar plot showing the number of datasets with unique runs found for each Virus group.
}
