% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_sum_hits_bar.R
\name{vh_sum_hits_bar}
\alias{vh_sum_hits_bar}
\title{Generate a bar plot showing the sum of hits for each virus family}
\usage{
vh_sum_hits_bar(vh_file)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunters hittables results.}
}
\value{
A list containing the generated bar plot and processed data
}
\description{
This function preprocesses virus data for plotting and generates a bar plot showing the sum of hits
for each virus family from the input dataset.
}
\details{
This function preprocesses virus data for plotting by calculating the sum of
hits for each virus family from the input dataset. It then generates a bar plot showing
the sum of hits for each virus family. Additionally, it returns the processed data
for further analysis.
}
