% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_sum_stat_evavlue_boxplot.R
\name{vh_sum_stat_evavlue_boxplot}
\alias{vh_sum_stat_evavlue_boxplot}
\title{Calculate summary statistics and identify hits below a specified E-value threshold for boxplots}
\usage{
vh_sum_stat_evavlue_boxplot(vh_file, cutoff)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter hittables results.}

\item{cutoff}{The significance cutoff value for E-values.}
}
\value{
A data frame containing summary statistics and the percentage of hits below the threshold for each virus group
}
\description{
This function calculates summary statistics, including median, quartiles, minimum, and maximum
-log10-transformed E-values, for each virus group from the VirusHunter hittables.
Additionally, it identifies hits below a specified E-value threshold and calculates
the percentage of hits below the threshold relative to the total hits for each virus group.
}
\details{
This function calculates summary statistics such as median, quartiles, minimum,
and maximum -log10-transformed E-values for each virus group from the VirusHunter hittables results.
It also identifies hits below a specified E-value threshold and calculates the percentage
of hits below the threshold relative to the total hits for each virus group.
}
\keyword{internal}
