% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vh_sumhitbar_preprocessing.R
\name{vh_sumhitbar_preprocessing}
\alias{vh_sumhitbar_preprocessing}
\title{Preprocess virus data for generating a bar plot of the sum of hits for each virus family}
\usage{
vh_sumhitbar_preprocessing(vh_file)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter hittables results}
}
\value{
A processed data frame with columns "best_query", "sum", "perc", "res", and "cyl"
}
\description{
This function preprocesses virus data for generating a bar plot showing
the sum of hits for each virus family from the input dataset.
}
\details{
This function preprocesses virus data for generating a bar plot showing the
sum of hits for each virus family from the input dataset. It calculates the sum of hits
for each virus family, computes the percentage of the sum relative to the total hits,
creates a string representation for visualization, and converts the "best_query" column
to a factor.
}
\author{
Sergej Ruff
}
\keyword{internal}
