% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhEvalBoxTable.R
\name{vhEvalBoxTable}
\alias{vhEvalBoxTable}
\title{vh_refeval_iden_boxplot: Create a formatted table using the gt package for vhEvalIdenBoxplot
with eval_vs_iden="evalue"}
\usage{
vhEvalBoxTable(summary_stats)
}
\arguments{
\item{summary_stats}{A data frame containing summary statistics}
}
\value{
A formatted gt table
}
\description{
This function takes the summary statistics output from vhEvalIdenBoxplot
and creates a formatted table using the gt package.Should only be used if the output stats were
generated with eval_vs_iden="evalue".
}
\details{
This function generates a table for the summary statistics output from the vhEvalIdenBoxplot
function.It formats summary statistics into a table using the gt package.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "evalue"
eval <- vhEvalIdenBoxplot(vh_file,eval_vs_iden="evalue",cut = 1e-5)

# generate table
eval_table <- vhEvalBoxTable(eval$summary_stats)

}
\seealso{
\code{\link{vhEvalIdenBoxplot}}
}
