% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhEvalIdenBoxplot.R
\name{vhEvalIdenBoxplot}
\alias{vhEvalIdenBoxplot}
\title{vhEvalIdenBoxplot: Generate boxplots comparing E-values or identity for each virus group}
\usage{
vhEvalIdenBoxplot(
  vh_file,
  eval_vs_iden = "evalue",
  cut = 1e-05,
  theme_choice = "minimal",
  x_label = NULL
)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter Hittable results.}

\item{eval_vs_iden}{Specifies whether to generate boxplots for E-values ("evalue")
or identity ("identity")}

\item{cut}{The significance cutoff value for E-values (default: 1e-5)}

\item{theme_choice}{(optional): A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "minimal".}

\item{x_label}{(optional): A character indicating new x label asigned by user. Default is NuLL
(NULl= "virus family query" )}
}
\value{
A list containing the generated boxplot, summary statistics, and outliers
}
\description{
This function generates boxplots comparing either E-values or identity
for each virus group from VirusHunter Hittables results.
}
\details{
This function generates boxplots comparing either E-values or identity for each virus
group from the VirusHunters Hittable.
It also calculates summary statistics and  identifies outliers for further analysis.
The user can specify whether to generate boxplots for E-values or identity by setting
the parameter \code{eval_vs_iden}.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "evalue"
eval <- vhEvalIdenBoxplot(vh_file,eval_vs_iden="evalue",cut = 1e-5)


# plot 2: plot boxplot for "identity"
identity <- vhEvalIdenBoxplot(vh_file,eval_vs_iden="identity",cut = 1e-5)

}
