% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{vhg_filter_belowthresholdboxplot}
\alias{vhg_filter_belowthresholdboxplot}
\title{internal function that extracts number of rows below threshold for vhgBoxplot function}
\usage{
vhg_filter_belowthresholdboxplot(vh_file, cut)
}
\arguments{
\item{vh_file}{vh_file}

\item{cut}{set cutoff in vhGBoxplot function}
}
\value{
filtered dataframe
}
\description{
internal function that extracts number of rows below threshold for vhgBoxplot function
}
\keyword{internal}
