% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VhgBoxplot.R
\name{VhgBoxplot}
\alias{VhgBoxplot}
\title{VhgBoxplot: Generate box plots comparing E-values,identity or contig length (Gatherer only) for each virus group}
\usage{
VhgBoxplot(
  file,
  x_column = "best_query",
  taxa_rank = "Family",
  y_column = "ViralRefSeq_E",
  contiglen_log10_scale = FALSE,
  cut = 1e-05,
  add_cutoff_line = TRUE,
  cut_colour = "#990000",
  reorder_criteria = "median",
  theme_choice = "linedraw",
  flip_coords = TRUE,
  add_mean_point = FALSE,
  mean_color = "white",
  mean_point_size = 2,
  title = "default",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = NULL,
  ylabel = NULL,
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  remove_group_labels = FALSE,
  legend_title = "Phylum",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10
)
}
\arguments{
\item{file}{A data frame containing VirusHunter or VirusGatherer Hittable results.}

\item{x_column}{A character specifying the column containing the groups (Default:"best_query").}

\item{taxa_rank}{(optional): When \code{x_column} is set to "ViralRefSeq_taxonomy", specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{y_column}{A character specifying the column containing the values to be compared. Currently "ViralRefSeq_ident",
"contig_len" (Column in Gatherer hittable) and "ViralRefSeq_E" are supported columns (Default:"ViralRefSeq_E").}

\item{contiglen_log10_scale}{(optional): When \code{y_column} is set to "contig_len", this parameter enables logarithmic scaling (log10) of the y-axis (TRUE). By default, this feature is disabled (FALSE).}

\item{cut}{(optional): The significance cutoff value for E-values (default: 1e-5).}

\item{add_cutoff_line}{(optional): whether to add a horizontal line based on \code{cut} for \code{"ViralRefSeq_E"} column (default: TRUE).}

\item{cut_colour}{(optional): The color for the significance cutoff line (default: "#990000").}

\item{reorder_criteria}{Character string specifying the criteria for reordering the x-axis ('max', 'min', 'median'(Default),'mean').
NULL sorts alphabetically.}

\item{theme_choice}{(optional):: A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "linedraw".}

\item{flip_coords}{(optional): Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{add_mean_point}{(optional): Logical indicating whether to add mean points to the boxplot. Default is FALSE.}

\item{mean_color}{(optional): change color of point indicating mean value in box plot (default: "white").}

\item{mean_point_size}{(optional): change size of point indicating mean value in box plot (default: 2)}

\item{title}{(optional): A character specifying the title of the plot. Default title is set based on y_column.}

\item{title_size}{(optional): Numeric specifying the size of the title text. Default is 16.}

\item{title_face}{(optional): A character specifying the font face for the title text. Default is "bold".}

\item{title_colour}{(optional): A character specifying the color for the title text. Default is "#2a475e".}

\item{subtitle}{(optional): A character specifying the subtitle of the plot. Default subtitle is set based on y_column.}

\item{subtitle_size}{(optional): Numeric specifying the size of the subtitle text. Default is 12.}

\item{subtitle_face}{(optional): A character specifying the font face for the subtitle text. Default is "bold".}

\item{subtitle_colour}{(optional): A character specifying the color for the subtitle text. Default is "#1b2838".}

\item{xlabel}{(optional): A character specifying the label for the x-axis. Default is "Virus found in query".}

\item{ylabel}{(optional): A character specifying the label for the y-axis. Default is set based on y_column.}

\item{axis_title_size}{(optional): Numeric specifying the size of the axis title text. Default is 12.}

\item{xtext_size}{(optional): Numeric specifying the size of the x-axis tick labels. Default is 10.}

\item{ytext_size}{(optional): Numeric specifying the size of the y-axis tick labels. Default is 10.}

\item{remove_group_labels}{(optional): If \code{TRUE}, the group labels will be removed; if \code{FALSE} or omitted, the labels will be displayed.}

\item{legend_title}{(optional): A character specifying the title for the legend. Default is "Phylum".}

\item{legend_position}{(optional): A character specifying the position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional): Numeric specifying the size of the legend title text. Default is 12.}

\item{legend_title_face}{(optional): A character specifying the font face for the legend title text. Default is "bold".}

\item{legend_text_size}{(optional): Numeric specifying the size of the legend text. Default is 10.}
}
\value{
A list containing:
\itemize{
\item the generated box plot.
\item summary statistics.
\item outliers ("ViralRefSeq_E" and "contig_len" only).
\item and rows_belowthres ("ViralRefSeq_E" only).
}
}
\description{
This function generates box plots comparing either E-values,identity or contig length (Gatherer only)
for each group from VirusHunter or VirusGatherer hittables results.
}
\details{
VhgBoxplot generates box plots comparing either E-values, identity, or contig length (Gatherer only) for each virus group from the VirusHunter or Gatherer Hittable.

The user can specify whether to generate boxplots for E-values, identity, or contig length (Gatherer only) by specifying the 'y_column'.
This means that 'VhgBoxplot' can generate three different types of boxplots.
By default, 'y_column' is set to "ViralRefSeq_E" and will plot the Reference E-Value on the y-axis.
Grouping on the x-axis is done by the 'x_column' argument. By default, the "best_query" will be used.

Additionally, the function calculates summary statistics and identifies outliers for further analysis ("ViralRefSeq_E" and "contig_len" only).
When 'y_column' is set to "ViralRefSeq_E", the output also includes 'rows_belowthres', which contains the hittable filtered for the rows below the threshold specified in the 'cut' argument.

The 'cut' argument is used differently depending on the 'y_column' value:
\itemize{
\item For 'y_column' set to "contig_len" or "ViralRefSeq_ident", the 'cut' argument filters the data to plot only the values with a "ViralRefSeq_E" below the specified threshold (default is 0.0001).
\item For 'y_column' set to "ViralRefSeq_E", the rows are not filtered. Instead, a horizontal line (h_line) is shown in the plot to indicate the cutoff value.
}

This allows the user to plot only the significant contig lengths and identities while also visualizing the number of non-significant and significant values for comparison.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
file <- ImportVirusTable(path)

# plot 1 for evalues
plot1 <- VhgBoxplot(file, x_column = "best_query", y_column = "ViralRefSeq_E")
plot1

# plot 2 for identity
plot2 <- VhgBoxplot(file, x_column = "best_query", y_column = "ViralRefSeq_ident")
plot2

# plot 3 custom arguments used
plot3 <- VhgBoxplot(file,
                  x_column = "best_query",
                  y_column = "ViralRefSeq_E",
                  theme_choice = "grey",
                  subtitle = "Custom subtitle: Identity for custom query",
                  xlabel = "Custom x-axis label: Custom query",
                  ylabel = "Custom y-axis label: Viral Reference Evalue in -log10 scale",
                  legend_position = "right")
plot3

# import gatherer files
path2 <- system.file("extdata", "virusgatherer.tsv", package = "Virusparies")
vg_file <- ImportVirusTable(path2)


# plot 4: Virusgatherer plot for ViralRefSeq_taxonomy agains contig length
plot5 <- VhgBoxplot(vg_file,x_column = "ViralRefSeq_taxonomy",y_column = "contig_len")
plot5



}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
