% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VhgEvalIdenBoxplot.R
\name{VhgEvalIdenBoxplot}
\alias{VhgEvalIdenBoxplot}
\title{vhEvalIdenBoxplot: Generate boxplots comparing E-values or identity for each virus group}
\usage{
VhgEvalIdenBoxplot(
  vh_file,
  x_column = "best_query",
  y_column = "ViralRefSeq_E",
  cut = 1e-05,
  cut_colour = "#990000",
  theme_choice = "minimal",
  flip_coords = TRUE,
  title = "default",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = NULL,
  ylabel = NULL,
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  legend_title = "virus family",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10
)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter or VirusGatherer Hittable results.}

\item{x_column}{A character specifying the column containing the groups (Default:"best_query").}

\item{y_column}{A character specifying the column containing the values to be compared. Currently "ViralRefSeq_ident" and
"ViralRefSeq_E" are supported columns (Default:"ViralRefSeq_E").}

\item{cut}{(optional) The significance cutoff value for E-values (default: 1e-5).}

\item{cut_colour}{(optional) The color for the significance cutoff line (default: "#990000").}

\item{theme_choice}{(optional): A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "minimal".}

\item{flip_coords}{(optional) Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional) A character specifying the title of the plot. Default title is set based on y_column.}

\item{title_size}{(optional) Numeric specifying the size of the title text. Default is 16.}

\item{title_face}{(optional) A character specifying the font face for the title text. Default is "bold".}

\item{title_colour}{(optional) A character specifying the color for the title text. Default is "#2a475e".}

\item{subtitle}{(optional) A character specifying the subtitle of the plot. Default subtitle is set based on y_column.}

\item{subtitle_size}{(optional) Numeric specifying the size of the subtitle text. Default is 12.}

\item{subtitle_face}{(optional) A character specifying the font face for the subtitle text. Default is "bold".}

\item{subtitle_colour}{(optional) A character specifying the color for the subtitle text. Default is "#1b2838".}

\item{xlabel}{(optional) A character specifying the label for the x-axis. Default is "Virus found in query".}

\item{ylabel}{(optional) A character specifying the label for the y-axis. Default is set based on y_column.}

\item{axis_title_size}{(optional) Numeric specifying the size of the axis title text. Default is 12.}

\item{xtext_size}{(optional) Numeric specifying the size of the x-axis tick labels. Default is 10.}

\item{ytext_size}{(optional) Numeric specifying the size of the y-axis tick labels. Default is 10.}

\item{legend_title}{(optional) A character specifying the title for the legend. Default is "virus family".}

\item{legend_position}{(optional) A character specifying the position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional) Numeric specifying the size of the legend title text. Default is 12.}

\item{legend_title_face}{(optional) A character specifying the font face for the legend title text. Default is "bold".}

\item{legend_text_size}{(optional) Numeric specifying the size of the legend text. Default is 10.}
}
\value{
A list containing the generated boxplot, summary statistics, and outliers
}
\description{
This function generates boxplots comparing either E-values or identity
for each group from VirusHunter or VirusGatherer Hittables results.
}
\details{
This function generates boxplots comparing either E-values or identity for each virus
group from the VirusHunter or Gatherer Hittable.
It also calculates summary statistics and  identifies outliers for further analysis.
The user can specify whether to generate boxplots for E-values or identity by setting
the parameter .
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1 for evalues
VhgEvalIdenBoxplot(vh_file, x_column = "best_query", y_column = "ViralRefSeq_E")

# plot 2 for identity
VhgEvalIdenBoxplot(vh_file, x_column = "best_query", y_column = "ViralRefSeq_ident")

# plot 3 custom arguments used
VhgEvalIdenBoxplot(vh_file,
                  x_column = "best_query",
                  y_column = "ViralRefSeq_E",
                  theme_choice = "grey",
                  subtitle = "Custom subtitle: Identity for custom query",
                  xlabel = "Custom x-axis label: Custom query",
                  ylabel = "Custom y-axis label: Viral Reference Evalue in -log10 scale",
                  legend_position = "right")

}
