% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VhgHist.R
\name{VhgHist}
\alias{VhgHist}
\title{VhgHist: Histogram and ridge density plots}
\usage{
VhgHist(
  file,
  plot_type = NULL,
  groupby = "best_query",
  taxa_rank = "Family",
  y_column = "ViralRefSeq_E",
  contiglen_log10_scale = FALSE,
  cut = 10000,
  alpha = 1,
  bins = NULL,
  binwidth = NULL,
  scale_ridge_plot = 1,
  ridge_bandwidth = NULL,
  add_jittered_points = FALSE,
  rel_min_height = 0,
  color_simple_hist = "gold",
  theme_choice = "linedraw",
  title = NULL,
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = NULL,
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = NULL,
  ylabel = NULL,
  axis_title_size = 12,
  xtext_size = 10,
  x_angle = NULL,
  ytext_size = 10,
  y_angle = NULL,
  remove_group_labels = FALSE,
  legend_title = "Legend",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  group_unwanted_phyla = NULL
)
}
\arguments{
\item{file}{A data frame containing VirusHunter or VirusGatherer hittable results.}

\item{plot_type}{Specifies the type of plot to generate:
\itemize{
\item \code{simple_histogram}: A basic histogram with no color differentiation.
\item \code{taxa_colored_histogram}: A histogram where bars are colored based on taxa/groups.
\item \code{ridge_plot}: A ridge plot showing density distributions for different taxa.
If \code{NULL}, defaults to \code{simple_histogram}.
}}

\item{groupby}{(optional): A character specifying the column containing the groups (default:"best_query").
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{taxa_rank}{(optional): When \code{groupby} is set to "ViralRefSeq_taxonomy", specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{y_column}{A character specifying the column containing the values to be compared. Currently "ViralRefSeq_ident",
"contig_len" (column in Gatherer hittable) and "ViralRefSeq_E" are supported columns (default:"ViralRefSeq_E").}

\item{contiglen_log10_scale}{(optional): When \code{y_column} is set to "contig_len", this parameter enables logarithmic scaling (log10) of the y-axis (TRUE). By default, this feature is disabled (FALSE).}

\item{cut}{(optional): The significance cutoff value for E-values (default: 10000).}

\item{alpha}{(optional): Alpha value for elements in plot (default: 1)}

\item{bins}{(optional): Number of bins. Overridden by binwidth. Defaults to 30.}

\item{binwidth}{(optional): The width of the bins. Can be specified as a numeric value or as a function that calculates width from unscaled x.}

\item{scale_ridge_plot}{(optional):  Determines the vertical scaling of the density curves in the ridge plot.}

\item{ridge_bandwidth}{(optional): Controls the smoothness of the density estimates in the ridge plot.
This parameter specifies the bandwidth of the kernel used for density estimation.}

\item{add_jittered_points}{(optional): Add jitter points to ridge plot (default: FALSE).}

\item{rel_min_height}{(optional): sets a percent cutoff relative to the highest point of any of the density curves to remove trailing tails
in ridge plot (default: 0)}

\item{color_simple_hist}{(optional): fill for \code{taxa_colored_histogram} (default: "gold").}

\item{theme_choice}{(optional): A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw" (default), and "test".
Append "_dotted" to any theme to add custom dotted grid lines (e.g., "classic_dotted").}

\item{title}{(optional): A character specifying the title of the plot.}

\item{title_size}{(optional): Numeric specifying the size of the title text (default: 16).}

\item{title_face}{(optional): A character specifying the font face for the title text (default: "bold").}

\item{title_colour}{(optional): A character specifying the color for the title text (default: "#2a475e").}

\item{subtitle}{(optional): A character specifying the subtitle of the plot.}

\item{subtitle_size}{(optional): Numeric specifying the size of the subtitle text(default: 12).}

\item{subtitle_face}{(optional): A character specifying the font face for the subtitle text (default: "bold").}

\item{subtitle_colour}{(optional): A character specifying the color for the subtitle text (default: "#1b2838").}

\item{xlabel}{(optional): A character specifying the label for the x-axis.}

\item{ylabel}{(optional): A character specifying the label for the y-axis.}

\item{axis_title_size}{(optional): Numeric specifying the size of the axis title text (default: 12).}

\item{xtext_size}{(optional): Numeric specifying the size of the x-axis tick labels (default: 10).}

\item{x_angle}{(optional): An integer specifying the angle (in degrees) for the x-axis text labels. Default is NULL, meaning no change.}

\item{ytext_size}{(optional): Numeric specifying the size of the y-axis tick labels (default: 10).}

\item{y_angle}{(optional): An integer specifying the angle (in degrees) for the y-axis text labels. Default is NULL, meaning no change.}

\item{remove_group_labels}{(optional): If \code{TRUE}, the group labels will be removed; if \code{FALSE} or omitted, the labels will be displayed.}

\item{legend_title}{(optional): A character specifying the title for the legend (default: "Legend").}

\item{legend_position}{(optional): A character specifying the position of the legend (default: "bottom").}

\item{legend_title_size}{(optional): Numeric specifying the size of the legend title text (default: 12).}

\item{legend_title_face}{(optional): A character specifying the font face for the legend title text (default: "bold").}

\item{legend_text_size}{(optional): Numeric specifying the size of the legend text (default: 10).}

\item{group_unwanted_phyla}{(optional): A character string specifying which group of viral phyla to retain in the analysis.
Valid values are:
\describe{
\item{"rna"}{Retain only the phyla specified for RNA viruses (\code{valid_phyla_rna}).}
\item{"smalldna"}{Retain only the phyla specified for small DNA viruses (\code{valid_phyla_smalldna}).}
\item{"largedna"}{Retain only the phyla specified for large DNA viruses (\code{valid_phyla_largedna}).}
}
All other phyla not in the specified group will be grouped into a single category:
"Non-RNA-virus" for \code{"rna"}, "Non-Small-DNA-Virus" for \code{"smalldna"}, or "Non-Large-DNA-Virus" for \code{"largedna"}.}
}
\value{
A plot obj.
}
\description{
VhgHist generates histogram and ridge density plots, enabling users to visualize the distribution and spread of their data.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
file <- ImportVirusTable(path)

# simple plot
VhgHist(file,plot_type = "simple_histogram",
y_column = ViralRefSeq_E,bins = 50,title = "Hist for -log10 of E-values",
ylabel = "Frequency",xlabel = "-log10 of viral reference E-values")

# plot filled according to  family
VhgHist(file,plot_type = "taxa_colored_histogram",y_column = ViralRefSeq_E,
bins = 50,title = "Taxa colored Hist for -log10 of E-values",
ylabel = "Frequency",xlabel = "-log10 of viral reference E-values",
groupby = ViralRefSeq_taxonomy,alpha = 0.7)

# ridge plot
VhgHist(file,plot_type = "ridge_plot",y_column = ViralRefSeq_E,
title = "Ridge plot for -log10 of E-values",ylabel = "Frequency",
xlabel = "-log10 of viral reference E-values",groupby = ViralRefSeq_taxonomy,
alpha = 0.7,scale_ridge_plot = 0.6)

}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
