% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VhgIdentityScatterPlot.R
\name{VhgIdentityScatterPlot}
\alias{VhgIdentityScatterPlot}
\title{VhgIdentityScatterPlot: Scatter Plot for Reference Identity vs -log10 of Reference E-value}
\usage{
VhgIdentityScatterPlot(
  vh_file,
  groupby = "best_query",
  taxa_rank = "Family",
  cutoff = 1e-05,
  theme_choice = "linedraw",
  cut_colour = "#990000",
  title = "Scatterplot of viral reference e-values and identity",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = NULL,
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Viral reference identity (\%)",
  ylabel = "-log10 of viral reference e-values",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  legend_title = "Phylum",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10
)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter or VirusGatherer Hittable results.}

\item{groupby}{(optional) A string indicating the column used for grouping the data points in the scatter plot.
The values in this column determine the color of each point in the scatter plot. Default is "best_query".
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{taxa_rank}{(optional) When \code{groupby} is set to "ViralRefSeq_taxonomy", specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{cutoff}{(optional) The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than the cutoff value in the ViralRefSeq_E column.}

\item{theme_choice}{(optional) A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test". Default is "linedraw".}

\item{cut_colour}{(optional) The color for the horizontal cutoff line. Default is "#990000".}

\item{title}{(optional) The title of the plot. Default is "Scatterplot of viral reference e-values and identity".}

\item{title_size}{(optional) The size of the title text. Default is 16.}

\item{title_face}{(optional) The face (bold, italic, etc.) of the title text. Default is "bold".}

\item{title_colour}{(optional) The color of the title text. Default is "#2a475e".}

\item{subtitle}{(optional) The subtitle of the plot. Default is NULL.}

\item{subtitle_size}{(optional) The size of the subtitle text. Default is 12.}

\item{subtitle_face}{(optional) The face (bold, italic, etc.) of the subtitle text. Default is "bold".}

\item{subtitle_colour}{(optional) The color of the subtitle text. Default is "#1b2838".}

\item{xlabel}{(optional) The label for the x-axis. Default is "Viral reference identity (\%)".}

\item{ylabel}{(optional) The label for the y-axis. Default is "-log10 of viral reference e-values".}

\item{axis_title_size}{(optional) The size of the axis titles. Default is 12.}

\item{xtext_size}{(optional) The size of the x-axis text. Default is 10.}

\item{ytext_size}{(optional) The size of the y-axis text. Default is 10.}

\item{legend_title}{(optional) The title of the legend. Default is "Phylum".}

\item{legend_position}{(optional) The position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional) The size of the legend title text. Default is 12.}

\item{legend_title_face}{(optional) The face (bold, italic, etc.) of the legend title text. Default is "bold".}

\item{legend_text_size}{(optional) The size of the legend text. Default is 10.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function creates a scatter plot of viral RefSeq identity versus
the negative logarithm (base 10) of viral RefSeq E-value. It colors the points
based on  best query  and adds a horizontal line representing the cutoff value.
}
\details{
VhgIdentityScatterPlot generates a scatter plot for Reference Identity versus -log10 of Reference E-value.
It accepts both VirusHunter and VirusGatherer Hittables as input.

A line indicates whether the observed values are above or below the cutoff specified by the 'cutoff' argument (default: 1e-5).
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- ImportVirusTable(path)


# Basic plot
plot <- VhgIdentityScatterPlot(vh_file,cutoff = 1e-5)

plot(plot$plot)

# Custom plot with additional arguments
custom_plot <- VhgIdentityScatterPlot(vh_file,
                                     cutoff = 1e-5,
                                     theme_choice = "dark",
                                     cut_colour = "blue",
                                     title = "Custom Scatter Plot",
                                     title_size = 18,
                                     title_face = "italic",
                                     title_colour = "darkred",
                                     subtitle = "Custom Subtitle",
                                     subtitle_size = 14,
                                     subtitle_face = "italic",
                                     subtitle_colour = "purple",
                                     xlabel = "Custom X Label",
                                     ylabel = "Custom Y Label",
                                     axis_title_size = 14,
                                     xtext_size = 12,
                                     ytext_size = 12,
                                     legend_title = "Custom Legend",
                                     legend_position = "top",
                                     legend_title_size = 14,
                                     legend_title_face = "italic",
                                     legend_text_size = 12)

plot(custom_plot$plot)

# import gatherer files
path2 <- system.file("extdata", "virusgatherer.tsv", package = "Virusparies")
vg_file <- ImportVirusTable(path2)

# vgplot: virusgatherer plot with ViralRefSeq_taxonomy as custom grouping
vgplot <- VhgIdentityScatterPlot(vg_file,groupby = "ViralRefSeq_taxonomy")
vgplot$plot

}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
