% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{VhgPreprocessTaxa}
\alias{VhgPreprocessTaxa}
\title{VhgPreprocessTaxa: preprocess ViralRefSeq_taxonomy elements}
\usage{
VhgPreprocessTaxa(vh_file, taxa_rank)
}
\arguments{
\item{vh_file}{VirusHunter or VirusGatherer hittable}

\item{taxa_rank}{Specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}
}
\value{
vh_file with preprocessed ViralRefSeq_taxonomy elements
}
\description{
VhgPreprocessTaxa: preprocess ViralRefSeq_taxonomy elements
}
\details{
Process the \code{ViralRefSeq_taxonomy} column.

Besides \code{best_query}, the user can utilize the \code{ViralRefSeq_taxonomy} column as \code{x_column} or \code{groupby} in plots.
This column needs preprocessing because it is too long and has too many unique elements for effective grouping.
The element containing the taxa suffix specified by the \code{taxa_rank} argument is used. NA values are replaced by "unclassified".

This function is used internally by every function that can use the \code{ViralRefSeq_taxonomy} column as input.
The user can also apply this function independently to process the taxonomy column and filter for the selected taxa rank in their data.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- ImportVirusTable(path)

vh_file_filtered <- VhgPreprocessTaxa(vh_file,"Family")

print("ViralRefSeq_taxonomy before processing:\n")
print(head(vh_file$ViralRefSeq_taxonomy,5))

print("ViralRefSeq_taxonomy after processing:\n")
print(head(vh_file_filtered$ViralRefSeq_taxonomy,5))



}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
