% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsBarplot.R
\name{VhgRunsBarplot}
\alias{VhgRunsBarplot}
\title{VhgRunsBarplot: Generate a bar plot showing the number of data sets with unique runs found for
each Virus group.}
\usage{
VhgRunsBarplot(
  file,
  groupby = "best_query",
  taxa_rank = "Family",
  cut = 1e-05,
  reorder_criteria = "max",
  theme_choice = "linedraw",
  flip_coords = TRUE,
  title = "Distribution of viral groups detected across query sequences",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Viral group",
  ylabel = "Number of datasets with hits for group",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  remove_group_labels = FALSE,
  legend_title = "Phylum",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  plot_text_size = 3.5,
  plot_text_position_dodge = 0.9,
  plot_text_hjust = -0.1,
  plot_text_vjust = 0.5,
  plot_text_colour = "black"
)
}
\arguments{
\item{file}{Data frame containing Virushunter or VirusGatherer hittables results.}

\item{groupby}{(optional) A string indicating the column used for grouping the data points in the plot.
"best_query" and "ViralRefSeq_taxonomy" can be used. Default is "best_query".
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{taxa_rank}{(optional) When \code{groupby} is set to "ViralRefSeq_taxonomy", specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{cut}{(optional) The significance cutoff value for E-values (default: 1e-5).
Removes rows in file with values larger than cutoff value in ViralRefSeq_E column.}

\item{reorder_criteria}{Character string specifying the criteria for reordering the x-axis ('max' (default), 'min').
NULL sorts alphabetically.}

\item{theme_choice}{(optional) A character indicating the ggplot2 theme to apply.
Options include "minimal", "classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test". Default is "linedraw".}

\item{flip_coords}{(optional) Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional) A character specifying the title of the plot. Default is "Distribution of viral groups detected across query sequences".}

\item{title_size}{(optional) The size of the plot title. Default is 16.}

\item{title_face}{(optional) The font face of the plot title. Default is "bold".}

\item{title_colour}{(optional) The color of the plot title. Default is "#2a475e".}

\item{subtitle}{(optional) A character specifying the subtitle of the plot.
Default is "default", which calculates the total number of data sets with hits and returns it as
"total number of data sets with hits: " followed by the calculated number. an empty string ("") removes the subtitle.}

\item{subtitle_size}{(optional) The size of the plot subtitle. Default is 12.}

\item{subtitle_face}{(optional) The font face of the plot subtitle. Default is "bold".}

\item{subtitle_colour}{(optional) The color of the plot subtitle. Default is "#1b2838".}

\item{xlabel}{(optional) A character specifying the label for the x-axis. Default is "Viral group".}

\item{ylabel}{(optional) A character specifying the label for the y-axis. Default is "Number of data sets with hits for group".}

\item{axis_title_size}{(optional) The size of axis titles. Default is 12.}

\item{xtext_size}{(optional) The size of x-axis text labels. Default is 10.}

\item{ytext_size}{(optional) The size of y-axis text labels. Default is 10.}

\item{remove_group_labels}{(optional): If \code{TRUE}, the group labels will be removed; if \code{FALSE} or omitted, the labels will be displayed.}

\item{legend_title}{(optional) A character specifying the title of the legend. Default is "Phylum".}

\item{legend_position}{(optional) The position of the legend. Default is "bottom"."none" removes legend.}

\item{legend_title_size}{(optional) The size of the legend title. Default is 12.}

\item{legend_title_face}{(optional) The font face of the legend title. Default is "bold".}

\item{legend_text_size}{(optional) The size of the legend text. Default is 10.}

\item{plot_text_size}{(optional) The size of the text labels added to the plot. Default is 3.5.}

\item{plot_text_position_dodge}{(optional) The degree of dodging for positioning text labels. Default is 0.9.}

\item{plot_text_hjust}{(optional) The horizontal justification of text labels. Default is -0.1.}

\item{plot_text_vjust}{(optional) The vertical justification of text labels. Default is 0.5.
It is recommended to change \code{vjust} when setting \code{flip_coords = FALSE}.}

\item{plot_text_colour}{(optional) The color of the text labels added to the plot. Default is "black".}
}
\value{
A list containing the bar plot and tabular data with information from the plot.
}
\description{
VhgRunsBarplot: Generate a bar plot showing the number of data sets with unique runs found for
each Virus group.
}
\details{
'VhgRunsBarplot' generates a bar plot showing the number of data sets with unique runs found for
each Virus group. It takes VirusHunter and VirusGatherer hittables as Input.

Only significant values below the threshold specified by the 'cut' argument (default: 1e-5) are included in the plot.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
file <- ImportVirusTable(path)

# plot 1: plot boxplot for "identity"
plot <- VhgRunsBarplot(file,cut = 1e-5)
plot

# return sample_run inside plot object
print(plot$sample_run)

# Plot 2: Customized plot with modified settings
plot_custom <- VhgRunsBarplot(
  file,
  cut = 1e-6, # Lower cutoff value
  theme_choice = "grey", # Classic theme
  title = "Customized Bar Plot", # Custom title
  subtitle = "test_subtitle", # No subtitle
  xlabel = "Custom X Label", # Custom x-axis label
  ylabel = "Custom Y Label", # Custom y-axis label
  legend_position = "top", # Legend position on top
  plot_text_size = 5, # Larger text label size
  plot_text_colour = "red" # Red text labels
)
plot_custom


# import gatherer files
path2 <- system.file("extdata", "virusgatherer.tsv", package = "Virusparies")
vg_file <- ImportVirusTable(path2)



}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
