% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsTable.R
\name{VhgRunsTable}
\alias{VhgRunsTable}
\title{VhgRunsTable: Generate a gt summary table of unique runs for each virus group}
\usage{
VhgRunsTable(
  vh_file,
  groupby = "best_query",
  cut = 1e-05,
  title = "Summary of unique runs by virus group",
  title_align = "left",
  names_ = NULL,
  align = "left",
  subtit = NULL,
  data_row.pad = 6,
  column_colour = "dodgerblue4",
  title_size = 26,
  subtitle_size = 14,
  title_weight = "bold",
  title_colour = "dodgerblue4",
  table_font_size = 14,
  cell_colour = "grey90",
  col_everyrow = FALSE
)
}
\arguments{
\item{vh_file}{A data frame containing the Virushunter Hittables Results}

\item{groupby}{(optional) A string indicating the column used for grouping the data.
"best_query" and "ViralRefSeq_taxonomy" can be used. Default is "best_query".
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{cut}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}

\item{title}{(optional): a title for the summary title.
Default is "Summary of unique runs by virus group"}

\item{title_align}{(optional): a character vector specifying the alignment of title (and subtile) text.
Possible values are "left", "center", or "right". Default is "left".}

\item{names_}{(optional): a vector of length 3 containing column names.
Default is c("Virus Group","Number of Unique SRA Runs","SRAs Found")}

\item{align}{(optional): a character vector specifying the alignment of text in the table columns.
Possible values are "left", "center", or "right". Default is "left".}

\item{subtit}{(optional): a character vector specifying the subtitle. Default is NULL.}

\item{data_row.pad}{(optional): numeric value specifying the row padding. Default is 6.}

\item{column_colour}{(optional): character specifying the background colour for the column header.
Default is dodgerblue4.}

\item{title_size}{(optional): numeric value specifying title size. Default is 26.}

\item{subtitle_size}{(optional): numeric value specifying subtitle size. Default is 14.}

\item{title_weight}{(optional): character or numeric value specifying title font weight.
The weight of the font can be modified thorough a text-based option such as "normal", "bold",
"lighter", "bolder", or, a numeric value between 1 and 1000, inclusive.Default value is "bold".}

\item{title_colour}{(optional): character specifying title colour. Default is dodgerblue4.}

\item{table_font_size}{(optional): numeric value specifying table font size. This will change font
size for the column header and for all values in each cell. Default is 14.}

\item{cell_colour}{(optional): character specifying cell colour.Default is grey90.}

\item{col_everyrow}{(optional): bool value specifying if every row or every second row of the table
should be filled with the colour from the cell_colour argument. col_everyrow = TRUE colours every
row and col_everyrow = FALSE colours every second row. Default is FALSE.}
}
\value{
A formatted gt table summarizing unique runs for each virus group
}
\description{
This function generates a summary table of unique runs for each virus group
based on the input dataset.
}
\details{
VhgRunsTable calculates the number of unique runs for each virus group from the input dataset.
It takes VirusHunter hittables as input and determines how many runs (SRA runs or local FastQ files) found a specific virus family.

A graphical table is returned with two columns. The first column contains the name of the virus group,
while the second column contains all IDs of SRA runs or local files that found that virus group.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# example 1: generate table with defaul arguments
table <- VhgRunsTable(vh_file,cut = 1e-5)

table

# example 2: generate table with custom arguments

table_2 <- VhgRunsTable(vh_file,title = "test",title_align="right",
names_ = c("column_1","column_2","column_3"),align = "right",subtit="subtitle")

table_2

}
\seealso{
\code{\link{VhgRunsBarplot}}
}
