% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhIdenFacetedScatterPlot.R
\name{vhIdenFacetedScatterPlot}
\alias{vhIdenFacetedScatterPlot}
\title{Create a Scatterplot of Viral RefSeq Identity vs. -log10 of Viral Reference e-value faceted by
the \code{best_query} column.}
\usage{
vhIdenFacetedScatterPlot(vh_file, cutoff = 1e-05)
}
\arguments{
\item{vh_file}{A data frame containing at least the columns \code{ViralRefSeq_E}, \code{ViralRefSeq_ident}, and \code{best_query}.}

\item{cutoff}{A numeric value representing the cutoff for the viral reference e-value. Points with \code{ViralRefSeq_E}
less than or equal to this value will be colored blue; otherwise, they will be colored red (default: 1e-5).}
}
\value{
A ggplot object representing the scatterplot.
}
\description{
This function generates a scatterplot of viral RefSeq identity versus -log10 of viral reference e-value
for each virus group in the \code{best_query} column. The points are colored based on whether the
e-value meets a specified cutoff and are faceted by the \code{best_query} column.
}
\details{
This function takes a data frame and a cutoff value as inputs, adds a new column to the
data frame indicating whether each \code{ViralRefSeq_E} value meets the cutoff, and then plots the data.
The plot includes:
\itemize{
\item Points colored based on whether they meet the cutoff condition.
\item Faceting by the \code{best_query} column.
}
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

vhIdenFacetedScatterPlot(vh_file,cutoff = 1e-5)


}
