% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhIdentityBoxTable.R
\name{vhIdentityBoxTable}
\alias{vhIdentityBoxTable}
\title{creatTableIdentityBox: Create a formatted table using the gt package for vhEvalIdenBoxplot
with eval_vs_iden="identity"}
\usage{
vhIdentityBoxTable(summary_stats_identity)
}
\arguments{
\item{summary_stats_identity}{A data frame containing summary statistics}
}
\value{
A gt table representing the summary statistics
}
\description{
This function takes the summary statistics output from vhEvalIdenBoxplot
and creates a formatted table using the gt package.Should only be used if the output stats were
generated with eval_vs_iden="identity".
}
\details{
This function generates a table for the summary statistics output from the vhEvalIdenBoxplot
function.It formats summary statistics into a table using the gt package.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "identity"
identity <- vhEvalIdenBoxplot(vh_file,eval_vs_iden="identity",cut = 1e-5)

# generate table
iden_table <- vhEvalBoxTable(identity$summary_stats)

iden_table

}
\seealso{
\code{\link{vhEvalIdenBoxplot}}
}
