% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhIdentityScatterPlot.R
\name{vhIdentityScatterPlot}
\alias{vhIdentityScatterPlot}
\title{vhIdentityScatterPlot: Scatter Plot for Reference Identity vs -log10 of Reference E-value}
\usage{
vhIdentityScatterPlot(vh_file, cutoff = 1e-05)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter Hittable results.}

\item{cutoff}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
This function creates a scatter plot of viral RefSeq identity versus
the negative logarithm (base 10) of viral RefSeq E-value. It colors the points
based on  best query  and adds a horizontal line representing the cutoff value.
}
