% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsBarplot.R
\name{vhRunsBarplot}
\alias{vhRunsBarplot}
\title{vhRunsBarplot: Generate a bar plot showing the number of datasets with unique runs found for
each Virus group.}
\usage{
vhRunsBarplot(vh_file, cut = 1e-05)
}
\arguments{
\item{vh_file}{Dataframe containing Virushunter hittables results}

\item{cut}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}
}
\value{
A list containing the bar plot and optionally the generated table and processed data
}
\description{
vhRunsBarplot: Generate a bar plot showing the number of datasets with unique runs found for
each Virus group.
}
