% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsTable.R
\name{vhRunsTable}
\alias{vhRunsTable}
\title{Generate a gt summary table of unique SRA runs for each virus group}
\usage{
vhRunsTable(vh_file, cut = 1e-05)
}
\arguments{
\item{vh_file}{A data frame containing the Virushunter Hittables Results}

\item{cut}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}
}
\value{
A formatted gt table summarizing unique runs for each virus group
}
\description{
This function generates a summary table of unique SRA runs for each virus group
based on the input dataset.
}
\details{
This function is an internal utility function used within the package.
It calculates the number of unique runs for each virus group from the input dataset.
}
\seealso{
\code{\link{vhRunsBarplot}}
}
