% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhSumHitsBarplot.R
\name{VhSumHitsBarplot}
\alias{VhSumHitsBarplot}
\title{VhSumHitsBarplot: Generate a bar plot showing the sum of hits for each virus family}
\usage{
VhSumHitsBarplot(
  vh_file,
  groupby = "best_query",
  taxa_rank = "Family",
  cut = 1e-05,
  reorder_criteria = "max",
  theme_choice = "linedraw",
  flip_coords = TRUE,
  title = "Distribution of hits for each virus group",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Viral group",
  ylabel = "Total number of hits",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  remove_group_labels = FALSE,
  legend_title = "Phylum",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  plot_text = 3,
  plot_text_size = 3.5,
  plot_text_position_dodge = 0.9,
  plot_text_hjust = -0.1,
  plot_text_vjust = 0.5,
  plot_text_colour = "black"
)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunters hittables results.}

\item{groupby}{(optional) A string indicating the column used for grouping the data points in the plot.
"best_query" and "ViralRefSeq_taxonomy" can be used. Default is "best_query".
Note: Gatherer hittables do not have a "best_query" column. Please provide an appropriate column for grouping.}

\item{taxa_rank}{(optional) When \code{groupby} is set to "ViralRefSeq_taxonomy", specify the taxonomic rank to group your data by.
Supported ranks are:
\itemize{
\item "Subphylum"
\item "Class"
\item "Subclass"
\item "Order"
\item "Suborder"
\item "Family" (default)
\item "Subfamily"
\item "Genus" (including Subgenus)
}}

\item{cut}{(optional) The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}

\item{reorder_criteria}{Character string specifying the criteria for reordering the x-axis ('max' (default), 'min').
NULL sorts alphabetically.}

\item{theme_choice}{(optional) A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "linedraw".}

\item{flip_coords}{(optional) Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional) A character specifying the title of the plot. Default is "Distribution of hits for each virus group".}

\item{title_size}{(optional) The size of the plot title. Default is 16.}

\item{title_face}{(optional) The font face of the plot title. Default is "bold".}

\item{title_colour}{(optional) The color of the plot title. Default is "#2a475e".}

\item{subtitle}{(optional) A character specifying the subtitle of the plot.
Default is "total number of hits: " followed by the calculated number.empty string ("") removes subttitle.}

\item{subtitle_size}{(optional) The size of the plot subtitle. Default is 12.}

\item{subtitle_face}{(optional) The font face of the plot subtitle. Default is "bold".}

\item{subtitle_colour}{(optional) The color of the plot subtitle. Default is "#1b2838".}

\item{xlabel}{(optional) A character specifying the label for the x-axis. Default is "Viral group".}

\item{ylabel}{(optional) A character specifying the label for the y-axis. Default is "Total number of hits".}

\item{axis_title_size}{(optional) The size of axis titles. Default is 12.}

\item{xtext_size}{(optional) The size of x-axis text labels. Default is 10.}

\item{ytext_size}{(optional) The size of y-axis text labels. Default is 10.}

\item{remove_group_labels}{(optional): If \code{TRUE}, the group labels will be removed; if \code{FALSE} or omitted, the labels will be displayed.}

\item{legend_title}{(optional) A character specifying the title of the legend. Default is "Phylum".}

\item{legend_position}{(optional) The position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional) The size of the legend title. Default is 12.}

\item{legend_title_face}{(optional) The font face of the legend title. Default is "bold".}

\item{legend_text_size}{(optional) The size of the legend text. Default is 10.}

\item{plot_text}{An index (0-3) to select the variable for text labels.
\itemize{
\item 0 = None.
\item 1 = Number of hits for each viral group.
\item 2 = Only the percentage.
\item 3 = Both (Default).
}}

\item{plot_text_size}{(optional) The size of the text labels added to the plot. Default is 3.5.}

\item{plot_text_position_dodge}{(optional) The degree of dodging for positioning text labels. Default is 0.9.}

\item{plot_text_hjust}{(optional) The horizontal justification of text labels. Default is -0.1.}

\item{plot_text_vjust}{(optional) The vertical justification of text labels. Default is 0.5.
It is recommended to change \code{vjust} when setting \code{flip_coords = FALSE}.}

\item{plot_text_colour}{(optional) The color of the text labels added to the plot. Default is "black".}
}
\value{
A list containing the generated bar plot and processed data.
}
\description{
This function preprocesses virus data for plotting and generates a bar plot showing the sum of hits/ contigs
for each virus family from the input dataset.
}
\details{
VhSumHitsBarplot preprocesses virus data for plotting by calculating the sum of
hits for each virus family from the input dataset (accepts only VirusHunter Hittables).
It then generates a bar plot showing the sum of hits/contigs for each virus family.
Additionally, it returns the processed data for further analysis.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- ImportVirusTable(path)

# plot 1: plot boxplot for "identity"
plot <- VhSumHitsBarplot(vh_file,cut = 1e-5)
plot

# return vh_group inside plot object
print(plot$vh_group)


# plot 2: Customized plot
plot2 <- VhSumHitsBarplot(
  vh_file,
  cut = 1e-5,
  title = "Customized Hits Plot",
  subtitle = "Total Hits with Significance Cutoff",
  xlabel = "Virus Family",
  ylabel = "Total Hits",
  legend_title = "Virus Families",
  plot_text_size = 4,
  plot_text_colour = "blue"
)

plot2

}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
