% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVirusGt.R
\name{exportVirusGt}
\alias{exportVirusGt}
\title{Export Graphical Tables}
\usage{
exportVirusGt(
  gtable,
  filename = "table.docx",
  path = NULL,
  create.dir = FALSE,
  ...
)
}
\arguments{
\item{gtable}{A graphical table object.}

\item{filename}{Name of the output file. Default is "table.docx".Make sure to provide an extension
compatible with the output types: .html, .tex, .ltx, .rtf, .docx. If a custom save function is provided,
the file extension is ignored.}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the
fully qualified file name. Defaults to the working directory.}

\item{create.dir}{Whether to create new directories if a non-existing directory is specified in
the filename or path (TRUE) or return an error (FALSE, default). If FALSE and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}

\item{...}{Pass any other options to the corresponding internal saving function.}
}
\value{
a message indicating that export was sucessful.
}
\description{
This function allows the user to export graphical tables in different formats.
}
\details{
export graphical tables (gt) generated by functions within the Virusparies package.

The exportVirusGt function, under the hood, utilizes the gt package for table manipulation and formatting.
For HTML output filenames with .html or .htm extensions, an HTML document is generated using the gt package.
Pass TRUE or FALSE to inline_css to include or exclude CSS styles inline (default is FALSE).
Additional options can be passed through .... For RTF output filenames with .rtf extension,
an RTF file is created. Use the page_numbering option to control page numbering (default is none).

For image files, use .png for PNG and .pdf for PDF. The gt package relies on Google Chrome installation
for PNG and PDF images. Pass options to webshot2::webshot() through .... Useful PNG options include zoom
(default: 2) and expand (default: 5).

For LaTeX output filenames with .tex, .ltx, or .rnw extensions, and .rtf for RTF, the corresponding
documents are generated. No additional options available.

For .docx output, requires rmarkdown package.

When 'create.dir' is set to TRUE, it generates a directory at the specified 'path' argument if the path doesn't already exist.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)



}
\author{
Sergej Ruff
}
