% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportVirusPlot.R
\name{ExportVirusPlot}
\alias{ExportVirusPlot}
\title{Export Virusparies Plots}
\usage{
ExportVirusPlot(
  file_name,
  plot = NULL,
  device = NULL,
  path = NULL,
  scale = 1,
  width = NA,
  height = NA,
  units = c("in", "cm", "mm", "px"),
  dpi = 300,
  limitsize = TRUE,
  bg = NULL,
  create.dir = FALSE,
  ...
)
}
\arguments{
\item{file_name}{Name of the output file.}

\item{plot}{The plot to be exported.}

\item{device}{The device used for output.
Can be one of "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", or "wmf" (windows only).
If NULL (default), the device is guessed based on the filename extension.}

\item{path}{The directory where the plot will be saved. Default is NULL (working directory).}

\item{scale}{A scaling factor. Default is 1.}

\item{width}{The width of the output file. Default is 800 pixels.}

\item{height}{The height of the output file. Default is 600 pixels.}

\item{units}{The units of the width and height parameters.
Can be one of "in", "cm", "mm", or "px". Default is "px".}

\item{dpi}{The resolution of the output device in dots per inch. Default is 300.}

\item{limitsize}{Whether to limit the size of the output file to the dimensions of the plot. Default is TRUE.}

\item{bg}{The background color of the plot. Default is NULL.}

\item{create.dir}{Whether to create the directory specified in the path if it does not exist. Default is FALSE.}

\item{...}{Additional arguments passed to.}
}
\value{
a message indicating that export was successful.
}
\description{
This function allows the user to export plots in different formats.
}
\details{
export plots generated by functions within the Virusparies package.

This function exports plots in various formats supported by Virusparies.
Supported devices include "eps", "ps", "tex", "pdf", "jpeg", "tiff", "png", "bmp", "svg", or "wmf" (Windows only).
When 'device' is set to NULL, the file extension in 'filename' is used to determine the device.

Depending on the plot, the final image might be cropped or truncated.
We recommend experimenting with height, width, and resolution.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# Basic plot
plot <- VhgIdentityScatterPlot(vh_file,cutoff = 1e-5)

\dontrun{

# first export
ExportVirusPlot(plot=plot,file_name="testplot.png",width=8,height=6,units="in")

# second export with device argument
ExportVirusPlot(plot=plot,file_name="testplot",width=8,height=6,units="in",device = "png")

}

}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
