% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importVirusTable.R
\name{importVirusTable}
\alias{importVirusTable}
\title{importVirusTable: Import VirusHunterGatherer hittables into R}
\usage{
importVirusTable(path)
}
\arguments{
\item{path}{A character string specifying the path to the VirusHunter or VirusGatherer hittable file.}
}
\value{
A dataframe containing the data from the VirusHunter or VirusGatherer hittable file.
}
\description{
This function imports VirusHunterGatherer hittables into R.
}
\details{
This function reads the VirusHunter or VirusGatherer hittable file specified by 'path'
into a dataframe in R. The file should be in tab-separated values (TSV) format.
The first row should contain column headers.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

print(head(vh_file))

# import gatherer files
path2 <- system.file("extdata", "virusgatherer.tsv", package = "Virusparies")
vg_file <- importVirusTable(path2)

print(head(vg_file))


}
\author{
Sergej Ruff
}
