% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-functions.R
\name{is_file_empty}
\alias{is_file_empty}
\title{Check if hittable is empty}
\usage{
is_file_empty(file)
}
\arguments{
\item{file}{a Virushuntegatherer hittable}
}
\description{
Check if hittable is empty
}
\details{
Internal helper function which checks if input is empty.objects with 0 observations (hittable,summarystats)
returns an error.
}
\keyword{internal}
