% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats_identity.R
\name{summary_stats_identity}
\alias{summary_stats_identity}
\title{Calculate summary statistics for identity}
\usage{
summary_stats_identity(vh_file)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunter hittables.}
}
\value{
A data frame containing summary statistics including median,
Q1 (1st quartile), Q3 (3rd quartile), minimum, and maximum identity values for each virus group
}
\description{
This function calculates summary statistics for identity based on the input virus data.
}
\details{
This function is an internal utility function used within the package.
It calculates summary statistics such as median, quartiles, minimum,
and maximum identity values for each virus group from the input dataset.
}
\keyword{internal}
