% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsBarplot.R
\name{vhRunsBarplot}
\alias{vhRunsBarplot}
\title{vhRunsBarplot: Generate a bar plot showing the number of datasets with unique runs found for
each Virus group.}
\usage{
vhRunsBarplot(
  vh_file,
  cut = 1e-05,
  theme_choice = "minimal",
  flip_coords = TRUE,
  title = "Number of datasets with hits found for each Virus group",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Virus family found in query",
  ylabel = "Number of datesets with hits",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  legend_title = "virus family",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  plot_text_size = 3.5,
  plot_text_position_dodge = 0.9,
  plot_text_hjust = -0.1,
  plot_text_vjust = 0.5,
  plot_text_colour = "black",
  colorblind = NULL
)
}
\arguments{
\item{vh_file}{Dataframe containing Virushunter hittables results.}

\item{cut}{(optional) The significance cutoff value for E-values (default: 1e-5).
Removes rows in vh_file with values larger than cutoff value in ViralRefSeq_E column.}

\item{theme_choice}{(optional) A character indicating the ggplot2 theme to apply.
Options include "minimal", "classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test". Default is "minimal".}

\item{flip_coords}{(optional) Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional) A character specifying the title of the plot. Default is "Number of datasets with hits found for each Virus group".}

\item{title_size}{(optional) The size of the plot title. Default is 16.}

\item{title_face}{(optional) The font face of the plot title. Default is "bold".}

\item{title_colour}{(optional) The color of the plot title. Default is "#2a475e".}

\item{subtitle}{(optional) A character specifying the subtitle of the plot.
Default is "default", which calculates the total number of datasets with hits and returns it as
"total number of datasets with hits: " followed by the calculated number. an empty string ("") removes the subtitle.}

\item{subtitle_size}{(optional) The size of the plot subtitle. Default is 12.}

\item{subtitle_face}{(optional) The font face of the plot subtitle. Default is "bold".}

\item{subtitle_colour}{(optional) The color of the plot subtitle. Default is "#1b2838".}

\item{xlabel}{(optional) A character specifying the label for the x-axis. Default is "Virus family found in query".}

\item{ylabel}{(optional) A character specifying the label for the y-axis. Default is "Number of datasets with hits".}

\item{axis_title_size}{(optional) The size of axis titles. Default is 12.}

\item{xtext_size}{(optional) The size of x-axis text labels. Default is 10.}

\item{ytext_size}{(optional) The size of y-axis text labels. Default is 10.}

\item{legend_title}{(optional) A character specifying the title of the legend. Default is "virus family".}

\item{legend_position}{(optional) The position of the legend. Default is "bottom"."none" removes legend.}

\item{legend_title_size}{(optional) The size of the legend title. Default is 12.}

\item{legend_title_face}{(optional) The font face of the legend title. Default is "bold".}

\item{legend_text_size}{(optional) The size of the legend text. Default is 10.}

\item{plot_text_size}{(optional) The size of the text labels added to the plot. Default is 3.5.}

\item{plot_text_position_dodge}{(optional) The degree of dodging for positioning text labels. Default is 0.9.}

\item{plot_text_hjust}{(optional) The horizontal justification of text labels. Default is -0.1.}

\item{plot_text_vjust}{(optional) The vertical justification of text labels. Default is 0.5.
It is recommended to change \code{vjust} when setting \code{flip_coords = FALSE}.}

\item{plot_text_colour}{(optional) The color of the text labels added to the plot. Default is "black".}

\item{colorblind}{(optional) A character specifying the type of color vision deficiency correction to apply.
Possible values are NULL (default), "deuteranope", "protanope", "desaturate", "enhanced", "enhanced.deuteranope",
"enhanced.protanope", and "enhanced.desaturate".
\itemize{
\item "deuteranope": Simulates deuteranopia (green-weakness).
\item "protanope": Simulates protanopia (red-weakness).
\item "desaturate": Converts colors to grayscale.
\item "enhanced": Applies enhanced color vision deficiency correction.
\item "enhanced.deuteranope": Enhanced correction for deuteranopia.
\item "enhanced.protanope": Enhanced correction for protanopia.
\item "enhanced.desaturate": Enhanced correction with desaturation.
}}
}
\value{
A list containing the bar plot and optionally the generated table and processed data
}
\description{
vhRunsBarplot: Generate a bar plot showing the number of datasets with unique runs found for
each Virus group.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "identity"
plot <- vhRunsBarplot(vh_file,cut = 1e-5)
plot

# return sample_run inside plot object
print(plot$sample_run)

# Plot 2: Customized plot with modified settings
plot_custom <- vhRunsBarplot(
  vh_file,
  cut = 1e-6, # Lower cutoff value
  theme_choice = "grey", # Classic theme
  title = "Customized Bar Plot", # Custom title
  subtitle = "test_subtitle", # No subtitle
  xlabel = "Custom X Label", # Custom x-axis label
  ylabel = "Custom Y Label", # Custom y-axis label
  legend_position = "top", # Legend position on top
  plot_text_size = 5, # Larger text label size
  plot_text_colour = "red" # Red text labels
)
plot_custom

}
