% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhRunsTable.R
\name{vhRunsTable}
\alias{vhRunsTable}
\title{Generate a gt summary table of unique runs for each virus group}
\usage{
vhRunsTable(
  vh_file,
  cut = 1e-05,
  title = "Summary Table of Unique Runs for Each Virus Group",
  title_align = "left",
  names_ = NULL,
  align = "left",
  subtit = NULL,
  data_row.pad = 6,
  column_colour = NULL,
  title_size = 26,
  subtitle_size = 14,
  title_weight = "bold",
  title_colour = NULL
)
}
\arguments{
\item{vh_file}{A data frame containing the Virushunter Hittables Results}

\item{cut}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}

\item{title}{(optional): a title for the summary title.
Default is "Summary Table of Unique Runs for Each Virus Group"}

\item{title_align}{(optional): a character vector specifying the alignment of title (and subttile) text.
Possible values are "left", "center", or "right". Default is "left".}

\item{names_}{(optional): a vector of length 3 containing column names.
Default is c("Virus Group","Number of Unique SRA Runs","SRAs Found")}

\item{align}{(optional): a character vector specifying the alignment of text in the table columns.
Possible values are "left", "center", or "right". Default is "left".}

\item{subtit}{(optional): a character vector specifying the subtitle. Default is NULL.}

\item{data_row.pad}{(optional): numeric value specifying the row padding. Default is 6.}

\item{column_colour}{(optional): character specifying the background colour for the column header.}

\item{title_size}{(optional): numeric value specifying title size. Defaulit is 26.}

\item{subtitle_size}{(optional): numeric value specifying subtitle size. Defaulit is 14.}

\item{title_weight}{(optional): character specifying title weight. Default value is "bold".}

\item{title_colour}{(optional): character specifying title colour. Default is NULL.}
}
\value{
A formatted gt table summarizing unique runs for each virus group
}
\description{
This function generates a summary table of unique runs for each virus group
based on the input dataset.
}
\details{
This function is an internal utility function used within the package.
It calculates the number of unique runs for each virus group from the input dataset.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# example 1: generate table with defaul arguments
table <- vhRunsTable(vh_file,cut = 1e-5)

table

# example 2: generate table with custom arguments

table_2 <- vhRunsTable(vh_file,title = "test",title_align="right",
names_ = c("column_1","column_2","column_3"),align = "right",subtit="subtitle")

table_2

}
\seealso{
\code{\link{vhRunsBarplot}}
}
