% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhSumHitsBarplot.R
\name{vhSumHitsBarplot}
\alias{vhSumHitsBarplot}
\title{Generate a bar plot showing the sum of hits for each virus family}
\usage{
vhSumHitsBarplot(
  vh_file,
  cut = 1e-05,
  theme_choice = "minimal",
  flip_coords = TRUE,
  title = "sum of hits for each family",
  title_size = 16,
  title_face = "bold",
  title_colour = "#2a475e",
  subtitle = "default",
  subtitle_size = 12,
  subtitle_face = "bold",
  subtitle_colour = "#1b2838",
  xlabel = "Virus family found in query",
  ylabel = "sum of hits",
  axis_title_size = 12,
  xtext_size = 10,
  ytext_size = 10,
  legend_title = "virus family",
  legend_position = "bottom",
  legend_title_size = 12,
  legend_title_face = "bold",
  legend_text_size = 10,
  plot_text_size = 3.5,
  plot_text_position_dodge = 0.9,
  plot_text_hjust = -0.1,
  plot_text_vjust = 0.5,
  plot_text_colour = "black",
  colorblind_support = FALSE,
  colormap = "viridis"
)
}
\arguments{
\item{vh_file}{A data frame containing VirusHunters hittables results.}

\item{cut}{(optional) The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}

\item{theme_choice}{(optional) A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "minimal".}

\item{flip_coords}{(optional) Logical indicating whether to flip the coordinates of the plot. Default is TRUE.}

\item{title}{(optional) A character specifying the title of the plot. Default is "sum of hits for each family".}

\item{title_size}{(optional) The size of the plot title. Default is 16.}

\item{title_face}{(optional) The font face of the plot title. Default is "bold".}

\item{title_colour}{(optional) The color of the plot title. Default is "#2a475e".}

\item{subtitle}{(optional) A character specifying the subtitle of the plot.
Default is "total number of hits: " followed by the calculated number.empty string ("") removes subttitle.}

\item{subtitle_size}{(optional) The size of the plot subtitle. Default is 12.}

\item{subtitle_face}{(optional) The font face of the plot subtitle. Default is "bold".}

\item{subtitle_colour}{(optional) The color of the plot subtitle. Default is "#1b2838".}

\item{xlabel}{(optional) A character specifying the label for the x-axis. Default is "Virus family found in query".}

\item{ylabel}{(optional) A character specifying the label for the y-axis. Default is "sum of hits".}

\item{axis_title_size}{(optional) The size of axis titles. Default is 12.}

\item{xtext_size}{(optional) The size of x-axis text labels. Default is 10.}

\item{ytext_size}{(optional) The size of y-axis text labels. Default is 10.}

\item{legend_title}{(optional) A character specifying the title of the legend. Default is "virus family".}

\item{legend_position}{(optional) The position of the legend. Default is "bottom".}

\item{legend_title_size}{(optional) The size of the legend title. Default is 12.}

\item{legend_title_face}{(optional) The font face of the legend title. Default is "bold".}

\item{legend_text_size}{(optional) The size of the legend text. Default is 10.}

\item{plot_text_size}{(optional) The size of the text labels added to the plot. Default is 3.5.}

\item{plot_text_position_dodge}{(optional) The degree of dodging for positioning text labels. Default is 0.9.}

\item{plot_text_hjust}{(optional) The horizontal justification of text labels. Default is -0.1.}

\item{plot_text_vjust}{(optional) The vertical justification of text labels. Default is 0.5.
It is recommended to change \code{vjust} when setting \code{flip_coords = FALSE}.}

\item{plot_text_colour}{(optional) The color of the text labels added to the plot. Default is "black".}

\item{colorblind_support}{(optional): Logical (TRUE or FALSE). If set to TRUE,
the function will use color scales that are more accessible for people with color vision deficiencies.
The default value is FALSE.}

\item{colormap}{(optional) Applies if \code{colorblind_support = TRUE}. A character string indicating the colormap option to use.
Default is "viridis". Eight options are available, derived from the Viridis package:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E")
\item "rocket" (or "F")
\item "mako" (or "G")
\item "turbo" (or "H")
}}
}
\value{
A list containing the generated bar plot and processed data
}
\description{
This function preprocesses virus data for plotting and generates a bar plot showing the sum of hits
for each virus family from the input dataset.
}
\details{
This function preprocesses virus data for plotting by calculating the sum of
hits for each virus family from the input dataset. It then generates a bar plot showing
the sum of hits for each virus family. Additionally, it returns the processed data
for further analysis.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "identity"
plot <- vhSumHitsBarplot(vh_file,cut = 1e-5)
plot

# return vh_group inside plot object
print(plot$vh_group)


# plot 2: Customized plot
plot2 <- vhSumHitsBarplot(
  vh_file,
  cut = 1e-5,
  title = "Customized Hits Plot",
  subtitle = "Total Hits with Significance Cutoff",
  xlabel = "Virus Family",
  ylabel = "Total Hits",
  legend_title = "Virus Families",
  plot_text_size = 4,
  plot_text_colour = "blue"
)

plot2

}
\seealso{
VirusHunterGatherer is available here: \url{https://github.com/lauberlab/VirusHunterGatherer}.
}
\author{
Sergej Ruff
}
