% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vhSumHitsBarplot.R
\name{vhSumHitsBarplot}
\alias{vhSumHitsBarplot}
\title{Generate a bar plot showing the sum of hits for each virus family}
\usage{
vhSumHitsBarplot(vh_file, cut = 1e-05, theme_choice = "minimal")
}
\arguments{
\item{vh_file}{A data frame containing VirusHunters hittables results.}

\item{cut}{The significance cutoff value for E-values (default: 1e-5). Removes rows in vh_file
with values larger than cutoff value in ViralRefSeq_E column.}

\item{theme_choice}{A character indicating the ggplot2 theme to apply. Options include "minimal",
"classic", "light", "dark", "void", "grey" (or "gray"), "bw", "linedraw", and "test".
Default is "minimal".}
}
\value{
A list containing the generated bar plot and processed data
}
\description{
This function preprocesses virus data for plotting and generates a bar plot showing the sum of hits
for each virus family from the input dataset.
}
\details{
This function preprocesses virus data for plotting by calculating the sum of
hits for each virus family from the input dataset. It then generates a bar plot showing
the sum of hits for each virus family. Additionally, it returns the processed data
for further analysis.
}
\examples{
path <- system.file("extdata", "virushunter.tsv", package = "Virusparies")
vh_file <- importVirusTable(path)

# plot 1: plot boxplot for "identity"
plot <- vhSumHitsBarplot(vh_file,cut = 1e-5)

# return vh_group inside plot object
print(plot$vh_group)

}
